ajout de la foreign key pour category

+++ compteurAppFrontend/src/app/pages/categories/categories.component.ts
@@ -4,6 +4,7 @@ import { Category } from 'src/models/category';
 import { Observable, map, startWith } from 'rxjs';
 import { Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
+import { FournisseurService } from 'src/app/_services/fournisseur.service';
 
 @Component({
   selector: 'app-categories',
@@ -17,7 +18,10 @@ export class CategoriesComponent implements OnInit, OnDestroy {
   selectedCategory!: Category;
   displayPopup = false;
 
-  constructor(private categoryService: CategoryService) { }
+  constructor(
+    private categoryService: CategoryService,
+    private fournisseurService: FournisseurService
+    ) { }
 
   ngOnInit(): void {
     this.loadCategories();
@@ -47,14 +51,32 @@ export class CategoriesComponent implements OnInit, OnDestroy {
   }
 
 
+  deleteMessageSucces = ""
+  deleteMessageError = ""
+
   deleteCategory(id: number | undefined): void {
-    this.categoryService.delete(id).pipe(takeUntil(this.destroy$)).subscribe(
-      () => {
-        this.closeDelete();
-        this.loadCategories();
+    this.fournisseurService.getUserByCategoryId(id).pipe(takeUntil(this.destroy$)).subscribe(
+      (data) => {
+        console.log(data);
+        if(data.length == 0) {
+          console.log(data)
+          this.categoryService.delete(id).pipe(takeUntil(this.destroy$)).subscribe(
+            () => this.loadCategories(),
+            error => console.error(error)
+          );
+          this.loadCategories();
+          this.deleteMessageSucces = "La catégorie a bien été supprimée"
+          this.deleteMessageError = ""
+          this.closeDelete();
+        } else {
+          this.closeDelete();
+          this.deleteMessageError = "Vous ne pouvez pas supprimer cette catégorie car elle contient des utilisateurs"
+          this.deleteMessageSucces = ""
+        }
       },
       error => console.error(error)
     );
+
   }
 
   categoryName = '';
